feat(settings): move editor preferences to global settings page
